Enable compiling non-QML API against Qt 5.15
Built with QtQuick disabled.

- C++17 required
- minor issues with QList vs QVector
- QMutexLocker templated vs not
- moc forward declaration handling

+++ src/animation/backend/animationclip.cpp
@@ -354,7 +354,7 @@ void AnimationClip::clearData()
 float AnimationClip::findDuration()
 {
     // Iterate over the contained fcurves and find the longest one
-    double tMax = 0.0;
+    float tMax = 0.f;
     for (const Channel &channel : qAsConst(m_channels)) {
         for (const ChannelComponent &channelComponent : qAsConst(channel.channelComponents)) {
             const float t = channelComponent.fcurve.endTime();
